Move function to subscription module
